[ANDROSDK-2013] move api logic to network layer
